test: move custom-dns-resolvers tests
